requirement.txt
appdirs==1.4.4
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
defusedxml==0.6.0
distlib==0.3.0
dj-database-url==0.5.0
Django==3.0.7
django-allauth==0.42.0
django-filter==2.2.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
filelock==3.0.12
gunicorn==20.0.4
idna==2.9
Markdown==3.2.2
oauthlib==3.1.0
psycopg2==2.8.5
psycopg2-binary==2.8.5
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
virtualenv==20.0.21
whitenoise==5.1.0
